style: forge fmt

+++ src/Reentrancy.sol
@@ -29,24 +29,24 @@ Here is how the functions were called
 */
 
 contract EtherStore {
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
     function deposit() public payable {
         balances[msg.sender] += msg.value;
     }
 
     function withdraw() public {
-        uint bal = balances[msg.sender];
+        uint256 bal = balances[msg.sender];
         require(bal > 0);
 
-        (bool sent, ) = msg.sender.call{value: bal}("");
+        (bool sent,) = msg.sender.call{value: bal}("");
         require(sent, "Failed to send Ether");
 
         balances[msg.sender] = 0;
     }
 
     // Helper function to check the balance of this contract
-    function getBalance() public view returns (uint) {
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
 }
@@ -72,8 +72,7 @@ contract Attack {
     }
 
     // Helper function to check the balance of this contract
-    function getBalance() public view returns (uint) {
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
 }
-
